requirements.txt
pyyaml
requests
dtweb
GitPython